Embargo emails and actions after login.
A script, check_embargoes.php, is provided.  If you run
this from cron it should detect embargoed jobs that have
expired and, if the user has provided an email address,
send them an email with relevant links to the job(s).
It will do this only once per job (and it sends information
on only 5 jobs per time).

If a user sends a URL containing form data (such as came from
the embargo email) but their session is no longer valid they
will be directed to the login screen as usual.  However on
successful login the original action will be executed.  E.g.
you enter the URL "portal?action=editjob&jid=10" but your session
has expired.  You are sent to the login screen, and after successful
login you are sent to "portal?action=editjob&jid=10".